ai/strands-quckstart/helloworld/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
python-dotenv>=1.0